Added better Error handling by deserializing proper error by endpoint
* Adds ResponseWithErr and ContentRespnoseWithErr
* Errors deserialize to the proper error type for the endpoint

The ResponseWithErr and ContentResponseWithErr were added because error_chain has to enumerate all error types and cannot contain generic errors. Since An error has to be of a concrete error type to be in a Result. We deserialize the response body to the error type expected and return the error when we know the exact error.